+++ src/pythonwrench/entrypoints/safe_rmdir.py
@@ -1,0 +1,96 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+import logging
+from argparse import ArgumentParser
+from pathlib import Path
+from typing import Union
+
+from pythonwrench.argparse import str_to_bool
+from pythonwrench.os import safe_rmdir
+
+logger = logging.getLogger(__name__)
+
+
+def print_safe_rmdir(
+    root: Union[str, Path],
+    *,
+    rm_root: bool = True,
+    error_on_non_empty_dir: bool = True,
+    followlinks: bool = False,
+    dry_run: bool = False,
+    verbose: int = 0,
+) -> None:
+    deleted, reviewed = safe_rmdir(
+        root=root,
+        rm_root=rm_root,
+        error_on_non_empty_dir=error_on_non_empty_dir,
+        followlinks=followlinks,
+        dry_run=dry_run,
+        verbose=verbose,
+    )
+    if dry_run:
+        msg = f"Dry run mode enabled. Here is the list of directories to delete ({len(deleted)}/{len(reviewed)}):"
+        print(msg)
+        for path in deleted:
+            print(f" - {path}")
+
+    elif verbose >= 1:
+        print(f"{len(deleted)} directories has been deleted.")
+
+
+def main_safe_rmdir() -> None:
+    parser = ArgumentParser()
+    parser.add_argument(
+        "root",
+        type=str,
+        help="Root directory path.",
+    )
+    parser.add_argument(
+        "--rm_root",
+        "--rm-root",
+        type=str_to_bool,
+        default=True,
+        help="If True, remove the root directory too if it is empty at the end. defaults to True.",
+    )
+    parser.add_argument(
+        "--error_on_non_empty_dir",
+        "--error-on-non-empty-dir",
+        type=str_to_bool,
+        default=True,
+        help="If True, raises a RuntimeError if a subdirectory contains at least 1 file. Otherwise it will ignore non-empty directories. defaults to True.",
+    )
+    parser.add_argument(
+        "--followlinks",
+        "--follow_links",
+        "--follow-links",
+        type=str_to_bool,
+        default=False,
+        help="Indicates whether or not symbolic links shound be followed. defaults to False.",
+    )
+    parser.add_argument(
+        "--dry_run",
+        "--dry-run",
+        type=str_to_bool,
+        default=False,
+        help="If True, does not remove any directory and just output the list of directories which could be deleted. defaults to False.",
+    )
+    parser.add_argument(
+        "--verbose",
+        type=int,
+        default=0,
+        help="Verbose level. defaults to 0.",
+    )
+    args = parser.parse_args()
+    print_safe_rmdir(
+        root=args.root,
+        rm_root=args.rm_root,
+        error_on_non_empty_dir=args.error_on_non_empty_dir,
+        followlinks=args.followlinks,
+        dry_run=args.dry_run,
+        verbose=args.verbose,
+    )
+
+
+if __name__ == "__main__":
+    main_safe_rmdir()